requirements.txt
requests
ttkthemes
reportlab
beautifulsoup4
PyPDF2
transformers
torch
tqdm
googlesearch-python
playsound